Promote a2alib out of experimental, remove hand-rolled a2a package
- Move experimental/a2alib → a2alib/ as part of the main dive module
  (no separate go.mod; a2a-go/v2 is now a direct dependency of dive)
- Add AgentCardProvider for per-request dynamic card resolution
- Add TestAgentCardProvider test (10 tests total, all passing)
- Remove experimental/a2a (hand-rolled A2A implementation superseded by a2alib)
- Remove examples/a2a_example; replace with examples/a2alib_example
- Add docs/guides/a2a.md; remove docs/guides/experimental/a2a.md
- Update CLAUDE.md package inventory
