+++ a2alib/executor.go
