+++ examples/go.mod
@@ -3,6 +3,7 @@ module github.com/deepnoodle-ai/dive/examples
 go 1.25.0
 
 require (
+	github.com/a2aproject/a2a-go/v2 v2.2.0
 	github.com/deepnoodle-ai/dive v1.4.0
 	github.com/deepnoodle-ai/dive/experimental/mcp v1.4.0
 	github.com/deepnoodle-ai/dive/otel v0.1.0
@@ -59,7 +60,9 @@ require (
 	go.opentelemetry.io/proto/otlp v1.10.0 // indirect
 	golang.org/x/crypto v0.49.0 // indirect
 	golang.org/x/image v0.38.0 // indirect
+	golang.org/x/mod v0.33.0 // indirect
 	golang.org/x/net v0.52.0 // indirect
+	golang.org/x/sync v0.20.0 // indirect
 	golang.org/x/sys v0.42.0 // indirect
 	golang.org/x/text v0.35.0 // indirect
 	google.golang.org/genai v1.51.0 // indirect
