--- experimental/a2a/a2a_test.go
@@ -1,1244 +1,0 @@
-package a2a_test
-
-import (
-	"context"
-	"encoding/json"
-	"fmt"
-	"net/http"
-	"net/http/httptest"
-	"strings"
-	"sync"
-	"sync/atomic"
-	"testing"
-
-	"github.com/deepnoodle-ai/dive"
-	"github.com/deepnoodle-ai/dive/experimental/a2a"
-	"github.com/deepnoodle-ai/dive/llm"
-	"github.com/deepnoodle-ai/wonton/assert"
-)
-
-// ---------------------------------------------------------------------------
-// Test doubles
-// ---------------------------------------------------------------------------
-
-type fakeLLM struct {
-	generate func(ctx context.Context, opts ...llm.Option) (*llm.Response, error)
-}
-
-func (f *fakeLLM) Name() string { return "fake-llm" }
-func (f *fakeLLM) Generate(ctx context.Context, opts ...llm.Option) (*llm.Response, error) {
-	return f.generate(ctx, opts...)
-}
-
-func textResponse(text string) *llm.Response {
-	return &llm.Response{
-		ID:         "resp_1",
-		Model:      "fake-model",
-		Role:       llm.Assistant,
-		Content:    []llm.Content{&llm.TextContent{Text: text}},
-		Type:       "message",
-		StopReason: "stop",
-		Usage:      llm.Usage{InputTokens: 5, OutputTokens: 5},
-	}
-}
-
-func toolCallResponse(toolName, callID string) *llm.Response {
-	return &llm.Response{
-		ID:    "resp_tool",
-		Model: "fake-model",
-		Role:  llm.Assistant,
-		Content: []llm.Content{&llm.ToolUseContent{
-			ID:    callID,
-			Name:  toolName,
-			Input: json.RawMessage(`{}`),
-		}},
-		Type:       "message",
-		StopReason: "tool_use",
-		Usage:      llm.Usage{InputTokens: 10, OutputTokens: 5},
-	}
-}
-
-func multiToolCallResponse(calls ...struct{ name, id string }) *llm.Response {
-	var content []llm.Content
-	for _, c := range calls {
-		content = append(content, &llm.ToolUseContent{
-			ID:    c.id,
-			Name:  c.name,
-			Input: json.RawMessage(`{}`),
-		})
-	}
-	return &llm.Response{
-		ID:         "resp_tool",
-		Model:      "fake-model",
-		Role:       llm.Assistant,
-		Content:    content,
-		Type:       "message",
-		StopReason: "tool_use",
-		Usage:      llm.Usage{InputTokens: 10, OutputTokens: 5},
-	}
-}
-
-func imageResponse(text string) *llm.Response {
-	return &llm.Response{
-		ID:    "resp_img",
-		Model: "fake-model",
-		Role:  llm.Assistant,
-		Content: []llm.Content{
-			&llm.TextContent{Text: text},
-			&llm.ImageContent{Source: &llm.ContentSource{
-				Type:      "url",
-				MediaType: "image/png",
-				URL:       "https://example.com/chart.png",
-			}},
-		},
-		Type:       "message",
-		StopReason: "stop",
-		Usage:      llm.Usage{InputTokens: 5, OutputTokens: 10},
-	}
-}
-
-type suspendingTool struct{}
-
-func (t *suspendingTool) Name() string                  { return "ask" }
-func (t *suspendingTool) Description() string           { return "Ask the human for input" }
-func (t *suspendingTool) Schema() *dive.Schema          { return nil }
-func (t *suspendingTool) Annotations() *dive.ToolAnnotations {
-	return &dive.ToolAnnotations{Title: "Ask"}
-}
-func (t *suspendingTool) Call(ctx context.Context, input any) (*dive.ToolResult, error) {
-	return dive.NewSuspendResult("Need your approval", map[string]any{"kind": "approval"}), nil
-}
-
-type confirmTool struct{}
-
-func (t *confirmTool) Name() string                         { return "confirm" }
-func (t *confirmTool) Description() string                  { return "Confirm with the human" }
-func (t *confirmTool) Schema() *dive.Schema                 { return nil }
-func (t *confirmTool) Annotations() *dive.ToolAnnotations   { return &dive.ToolAnnotations{Title: "Confirm"} }
-func (t *confirmTool) Call(ctx context.Context, input any) (*dive.ToolResult, error) {
-	return dive.NewSuspendResult("Please confirm", nil), nil
-}
-
-type blockingLLM struct {
-	started chan struct{}
-	release chan struct{}
-	result  *llm.Response
-}
-
-func (f *blockingLLM) Name() string { return "blocking-llm" }
-func (f *blockingLLM) Generate(ctx context.Context, opts ...llm.Option) (*llm.Response, error) {
-	close(f.started)
-	select {
-	case <-ctx.Done():
-		return nil, ctx.Err()
-	case <-f.release:
-		return f.result, nil
-	}
-}
-
-// ---------------------------------------------------------------------------
-// Types / encoding
-// ---------------------------------------------------------------------------
-
-func TestTaskStateIsTerminal(t *testing.T) {
-	terminal := []a2a.TaskState{
-		a2a.TaskStateCompleted,
-		a2a.TaskStateCanceled,
-		a2a.TaskStateFailed,
-		a2a.TaskStateRejected,
-	}
-	for _, s := range terminal {
-		assert.True(t, s.IsTerminal())
-	}
-	nonTerminal := []a2a.TaskState{
-		a2a.TaskStateSubmitted,
-		a2a.TaskStateWorking,
-		a2a.TaskStateInputRequired,
-		a2a.TaskStateAuthRequired,
-	}
-	for _, s := range nonTerminal {
-		assert.False(t, s.IsTerminal())
-	}
-}
-
-func TestMessageJSONOmitsNullParts(t *testing.T) {
-	msg := a2a.Message{MessageID: "m1", Role: a2a.RoleUser}
-	b, err := json.Marshal(msg)
-	assert.NoError(t, err)
-	assert.True(t, strings.Contains(string(b), `"parts":[]`))
-}
-
-func TestSendMessageParamsValidate(t *testing.T) {
-	valid := &a2a.SendMessageParams{
-		Message: &a2a.Message{
-			MessageID: "m1",
-			Role:      a2a.RoleUser,
-			Parts:     []a2a.Part{a2a.NewTextPart("hello")},
-		},
-	}
-	assert.NoError(t, valid.Validate())
-
-	cases := []*a2a.SendMessageParams{
-		nil,
-		{},
-		{Message: &a2a.Message{Role: a2a.RoleUser, Parts: []a2a.Part{a2a.NewTextPart("x")}}},
-		{Message: &a2a.Message{MessageID: "m", Parts: []a2a.Part{a2a.NewTextPart("x")}}},
-		{Message: &a2a.Message{MessageID: "m", Role: a2a.RoleAgent, Parts: []a2a.Part{a2a.NewTextPart("x")}}},
-		{Message: &a2a.Message{MessageID: "m", Role: a2a.RoleUser}},
-	}
-	for i, c := range cases {
-		err := c.Validate()
-		assert.Error(t, err)
-		_ = i
-	}
-}
-
-// ---------------------------------------------------------------------------
-// Server: well-known agent card
-// ---------------------------------------------------------------------------
-
-func buildAgent(t *testing.T, model llm.LLM, tools ...dive.Tool) *dive.Agent {
-	t.Helper()
-	agent, err := dive.NewAgent(dive.AgentOptions{
-		Name:  "Research Assistant",
-		Model: model,
-		Tools: tools,
-	})
-	assert.NoError(t, err)
-	return agent
-}
-
-func buildParallelAgent(t *testing.T, model llm.LLM, tools ...dive.Tool) *dive.Agent {
-	t.Helper()
-	agent, err := dive.NewAgent(dive.AgentOptions{
-		Name:                  "Research Assistant",
-		Model:                 model,
-		Tools:                 tools,
-		ParallelToolExecution: true,
-	})
-	assert.NoError(t, err)
-	return agent
-}
-
-func TestServerAgentCard(t *testing.T) {
-	model := &fakeLLM{generate: func(ctx context.Context, opts ...llm.Option) (*llm.Response, error) {
-		return textResponse("ok"), nil
-	}}
-	agent := buildAgent(t, model)
-
-	server, err := a2a.NewServer(a2a.ServerOptions{
-		Agent:   agent,
-		BaseURL: "https://agent.example.com",
-	})
-	assert.NoError(t, err)
-
-	ts := httptest.NewServer(server.Handler())
-	defer ts.Close()
-
-	resp, err := http.Get(ts.URL + "/.well-known/agent.json")
-	assert.NoError(t, err)
-	defer resp.Body.Close()
-	assert.Equal(t, resp.StatusCode, http.StatusOK)
-
-	var card a2a.AgentCard
-	assert.NoError(t, json.NewDecoder(resp.Body).Decode(&card))
-	assert.Equal(t, card.Name, "Research Assistant")
-	assert.True(t, card.Capabilities.Streaming)
-	assert.True(t, len(card.SupportedInterfaces) > 0)
-	assert.True(t, strings.HasPrefix(card.SupportedInterfaces[0].URL, "https://agent.example.com"))
-}
-
-// ---------------------------------------------------------------------------
-// Server: message/send happy path
-// ---------------------------------------------------------------------------
-
-func TestServerMessageSendCompletion(t *testing.T) {
-	model := &fakeLLM{generate: func(ctx context.Context, opts ...llm.Option) (*llm.Response, error) {
-		return textResponse("Paris is the capital of France."), nil
-	}}
-	agent := buildAgent(t, model)
-
-	server, err := a2a.NewServer(a2a.ServerOptions{Agent: agent})
-	assert.NoError(t, err)
-
-	ts := httptest.NewServer(server.Handler())
-	defer ts.Close()
-
-	client, err := a2a.NewClient(a2a.ClientOptions{Endpoint: ts.URL + "/"})
-	assert.NoError(t, err)
-
-	task, err := client.SendMessage(context.Background(), &a2a.Message{
-		MessageID: "m1",
-		Role:      a2a.RoleUser,
-		Parts:     []a2a.Part{a2a.NewTextPart("What is the capital of France?")},
-	}, nil)
-	assert.NoError(t, err)
-	assert.Equal(t, task.Status.State, a2a.TaskStateCompleted)
-	assert.Equal(t, a2a.ResponseText(task), "Paris is the capital of France.")
-	assert.NotEqual(t, task.ID, "")
-	assert.NotEqual(t, task.ContextID, "")
-}
-
-// ---------------------------------------------------------------------------
-// Server: suspend → input-required → resume
-// ---------------------------------------------------------------------------
-
-func TestServerSuspendMapsToInputRequired(t *testing.T) {
-	var callNum atomic.Int32
-	model := &fakeLLM{generate: func(ctx context.Context, opts ...llm.Option) (*llm.Response, error) {
-		n := callNum.Add(1)
-		switch n {
-		case 1:
-			return toolCallResponse("ask", "call_1"), nil
-		case 2:
-			return textResponse("Thanks, I have what I need."), nil
-		}
-		return textResponse("unexpected"), nil
-	}}
-	agent := buildAgent(t, model, &suspendingTool{})
-
-	server, err := a2a.NewServer(a2a.ServerOptions{Agent: agent})
-	assert.NoError(t, err)
-
-	ts := httptest.NewServer(server.Handler())
-	defer ts.Close()
-
-	client, err := a2a.NewClient(a2a.ClientOptions{Endpoint: ts.URL + "/"})
-	assert.NoError(t, err)
-
-	task, err := client.SendMessage(context.Background(), &a2a.Message{
-		MessageID: "m1",
-		Role:      a2a.RoleUser,
-		Parts:     []a2a.Part{a2a.NewTextPart("Please proceed")},
-	}, nil)
-	assert.NoError(t, err)
-	assert.Equal(t, task.Status.State, a2a.TaskStateInputRequired)
-	assert.NotNil(t, task.Status.Message)
-	assert.Equal(t, task.Status.Message.TextContent(), "Need your approval")
-
-	// Resume by sending a new message targeting the same task.
-	resumed, err := client.SendMessage(context.Background(), &a2a.Message{
-		MessageID: "m2",
-		Role:      a2a.RoleUser,
-		Parts:     []a2a.Part{a2a.NewTextPart("approved")},
-		TaskID:    task.ID,
-		ContextID: task.ContextID,
-	}, nil)
-	assert.NoError(t, err)
-	assert.Equal(t, resumed.Status.State, a2a.TaskStateCompleted)
-	assert.Equal(t, a2a.ResponseText(resumed), "Thanks, I have what I need.")
-}
-
-// ---------------------------------------------------------------------------
-// Server: tasks/get and tasks/cancel
-// ---------------------------------------------------------------------------
-
-func TestServerTasksGetAndCancel(t *testing.T) {
-	var callNum atomic.Int32
-	model := &fakeLLM{generate: func(ctx context.Context, opts ...llm.Option) (*llm.Response, error) {
-		n := callNum.Add(1)
-		if n == 1 {
-			return toolCallResponse("ask", "call_1"), nil
-		}
-		return textResponse("done"), nil
-	}}
-	agent := buildAgent(t, model, &suspendingTool{})
-
-	server, err := a2a.NewServer(a2a.ServerOptions{Agent: agent})
-	assert.NoError(t, err)
-
-	ts := httptest.NewServer(server.Handler())
-	defer ts.Close()
-
-	client, err := a2a.NewClient(a2a.ClientOptions{Endpoint: ts.URL + "/"})
-	assert.NoError(t, err)
-
-	task, err := client.SendMessage(context.Background(), &a2a.Message{
-		MessageID: "m1",
-		Role:      a2a.RoleUser,
-		Parts:     []a2a.Part{a2a.NewTextPart("Please proceed")},
-	}, nil)
-	assert.NoError(t, err)
-	assert.Equal(t, task.Status.State, a2a.TaskStateInputRequired)
-
-	got, err := client.GetTask(context.Background(), task.ID)
-	assert.NoError(t, err)
-	assert.Equal(t, got.ID, task.ID)
-	assert.Equal(t, got.Status.State, a2a.TaskStateInputRequired)
-
-	canceled, err := client.CancelTask(context.Background(), task.ID)
-	assert.NoError(t, err)
-	assert.Equal(t, canceled.Status.State, a2a.TaskStateCanceled)
-
-	// Cancelling again should fail with TaskNotCancelable.
-	_, err = client.CancelTask(context.Background(), task.ID)
-	assert.Error(t, err)
-	rpcErr, ok := err.(*a2a.RPCError)
-	assert.True(t, ok)
-	assert.Equal(t, rpcErr.Code, a2a.ErrorCodeTaskNotCancelable)
-}
-
-// ---------------------------------------------------------------------------
-// Server: tasks/get for unknown id
-// ---------------------------------------------------------------------------
-
-func TestServerTasksGetUnknown(t *testing.T) {
-	model := &fakeLLM{generate: func(ctx context.Context, opts ...llm.Option) (*llm.Response, error) {
-		return textResponse("ok"), nil
-	}}
-	agent := buildAgent(t, model)
-	server, err := a2a.NewServer(a2a.ServerOptions{Agent: agent})
-	assert.NoError(t, err)
-	ts := httptest.NewServer(server.Handler())
-	defer ts.Close()
-
-	client, err := a2a.NewClient(a2a.ClientOptions{Endpoint: ts.URL + "/"})
-	assert.NoError(t, err)
-
-	_, err = client.GetTask(context.Background(), "does-not-exist")
-	assert.Error(t, err)
-	rpcErr, ok := err.(*a2a.RPCError)
-	assert.True(t, ok)
-	assert.Equal(t, rpcErr.Code, a2a.ErrorCodeTaskNotFound)
-}
-
-// ---------------------------------------------------------------------------
-// Server: streaming
-// ---------------------------------------------------------------------------
-
-func TestServerMessageStream(t *testing.T) {
-	model := &fakeLLM{generate: func(ctx context.Context, opts ...llm.Option) (*llm.Response, error) {
-		return textResponse("streamed hello"), nil
-	}}
-	agent := buildAgent(t, model)
-	server, err := a2a.NewServer(a2a.ServerOptions{Agent: agent})
-	assert.NoError(t, err)
-	ts := httptest.NewServer(server.Handler())
-	defer ts.Close()
-
-	client, err := a2a.NewClient(a2a.ClientOptions{Endpoint: ts.URL + "/"})
-	assert.NoError(t, err)
-
-	var (
-		statusEvents   []*a2a.TaskStatusUpdateEvent
-		artifactEvents []*a2a.TaskArtifactUpdateEvent
-	)
-	err = client.StreamMessage(context.Background(), &a2a.Message{
-		MessageID: "m1",
-		Role:      a2a.RoleUser,
-		Parts:     []a2a.Part{a2a.NewTextPart("stream please")},
-	}, nil, func(ev *a2a.StreamEvent) error {
-		switch {
-		case ev.StatusUpdate != nil:
-			statusEvents = append(statusEvents, ev.StatusUpdate)
-		case ev.ArtifactUpdate != nil:
-			artifactEvents = append(artifactEvents, ev.ArtifactUpdate)
-		}
-		return nil
-	})
-	assert.NoError(t, err)
-	assert.True(t, len(statusEvents) > 0)
-	final := statusEvents[len(statusEvents)-1]
-	assert.Equal(t, final.Status.State, a2a.TaskStateCompleted)
-	assert.Len(t, artifactEvents, 1)
-	assert.Equal(t, artifactEvents[0].Artifact.Parts[0].Text, "streamed hello")
-}
-
-// ---------------------------------------------------------------------------
-// RemoteAgent wrapper
-// ---------------------------------------------------------------------------
-
-func TestRemoteAgentSendText(t *testing.T) {
-	model := &fakeLLM{generate: func(ctx context.Context, opts ...llm.Option) (*llm.Response, error) {
-		return textResponse("remote response"), nil
-	}}
-	agent := buildAgent(t, model)
-	server, err := a2a.NewServer(a2a.ServerOptions{Agent: agent})
-	assert.NoError(t, err)
-	ts := httptest.NewServer(server.Handler())
-	defer ts.Close()
-
-	client, err := a2a.NewClient(a2a.ClientOptions{Endpoint: ts.URL + "/"})
-	assert.NoError(t, err)
-	remote := a2a.NewRemoteAgent(client)
-
-	task, err := remote.SendText(context.Background(), "ping")
-	assert.NoError(t, err)
-	assert.Equal(t, a2a.ResponseText(task), "remote response")
-	assert.NotEqual(t, remote.ContextID(), "")
-
-	// Second call should reuse the contextId.
-	prevCtxID := remote.ContextID()
-	task2, err := remote.SendText(context.Background(), "again")
-	assert.NoError(t, err)
-	assert.Equal(t, task2.ContextID, prevCtxID)
-}
-
-// ---------------------------------------------------------------------------
-// JSON-RPC error cases
-// ---------------------------------------------------------------------------
-
-func TestServerInvalidMethod(t *testing.T) {
-	model := &fakeLLM{generate: func(ctx context.Context, opts ...llm.Option) (*llm.Response, error) {
-		return textResponse("ok"), nil
-	}}
-	agent := buildAgent(t, model)
-	server, err := a2a.NewServer(a2a.ServerOptions{Agent: agent})
-	assert.NoError(t, err)
-	ts := httptest.NewServer(server.Handler())
-	defer ts.Close()
-
-	body := `{"jsonrpc":"2.0","id":1,"method":"mystery/do","params":{}}`
-	resp, err := http.Post(ts.URL+"/", "application/json", strings.NewReader(body))
-	assert.NoError(t, err)
-	defer resp.Body.Close()
-	var env struct {
-		Error *a2a.RPCError `json:"error"`
-	}
-	assert.NoError(t, json.NewDecoder(resp.Body).Decode(&env))
-	assert.NotNil(t, env.Error)
-	assert.Equal(t, env.Error.Code, a2a.ErrorCodeMethodNotFound)
-}
-
-// ---------------------------------------------------------------------------
-// Server: known-but-unimplemented methods report UnsupportedOperation
-// ---------------------------------------------------------------------------
-
-func TestServerUnsupportedMethodsReportCorrectCode(t *testing.T) {
-	model := &fakeLLM{generate: func(ctx context.Context, opts ...llm.Option) (*llm.Response, error) {
-		return textResponse("ok"), nil
-	}}
-	agent := buildAgent(t, model)
-	server, err := a2a.NewServer(a2a.ServerOptions{Agent: agent})
-	assert.NoError(t, err)
-	ts := httptest.NewServer(server.Handler())
-	defer ts.Close()
-
-	methods := []string{
-		a2a.MethodTasksResubscribe,
-		a2a.MethodTasksPushNotifConfigSet,
-		a2a.MethodTasksPushNotifConfigGet,
-		a2a.MethodTasksPushNotifConfigList,
-		a2a.MethodTasksPushNotifConfigDelete,
-		a2a.MethodAgentExtendedCard,
-	}
-	for _, method := range methods {
-		body := `{"jsonrpc":"2.0","id":1,"method":"` + method + `","params":{}}`
-		resp, err := http.Post(ts.URL+"/", "application/json", strings.NewReader(body))
-		assert.NoError(t, err)
-		var env struct {
-			Error *a2a.RPCError `json:"error"`
-		}
-		assert.NoError(t, json.NewDecoder(resp.Body).Decode(&env))
-		resp.Body.Close()
-		assert.NotNil(t, env.Error)
-		assert.Equal(t, env.Error.Code, a2a.ErrorCodeUnsupportedOperation)
-		assert.True(t, strings.Contains(env.Error.Message, method))
-	}
-}
-
-// ---------------------------------------------------------------------------
-// Server: mixed content parts project into typed LLM content
-// ---------------------------------------------------------------------------
-
-func TestServerMixedPartsProjectToTypedContent(t *testing.T) {
-	var textContent string
-	var hasDocument bool
-	var docURI string
-	model := &fakeLLM{generate: func(ctx context.Context, opts ...llm.Option) (*llm.Response, error) {
-		cfg := &llm.Config{}
-		cfg.Apply(opts...)
-		for _, msg := range cfg.Messages {
-			if msg.Role != llm.User {
-				continue
-			}
-			for _, c := range msg.Content {
-				switch v := c.(type) {
-				case *llm.TextContent:
-					textContent += v.Text
-				case *llm.DocumentContent:
-					hasDocument = true
-					if v.Source != nil {
-						docURI = v.Source.URL
-					}
-				}
-			}
-		}
-		return textResponse("got it"), nil
-	}}
-	agent := buildAgent(t, model)
-	server, err := a2a.NewServer(a2a.ServerOptions{Agent: agent})
-	assert.NoError(t, err)
-	ts := httptest.NewServer(server.Handler())
-	defer ts.Close()
-
-	client, err := a2a.NewClient(a2a.ClientOptions{Endpoint: ts.URL + "/"})
-	assert.NoError(t, err)
-
-	task, err := client.SendMessage(context.Background(), &a2a.Message{
-		MessageID: "m1",
-		Role:      a2a.RoleUser,
-		Parts: []a2a.Part{
-			a2a.NewTextPart("Summarize the attached order."),
-			a2a.NewDataPart(map[string]any{"orderId": "ABC-123", "total": 4299}),
-			{URL: "https://example.com/invoice.pdf", MediaType: "application/pdf", Filename: "invoice.pdf"},
-		},
-	}, nil)
-	assert.NoError(t, err)
-	assert.Equal(t, task.Status.State, a2a.TaskStateCompleted)
-	assert.True(t, strings.Contains(textContent, "Summarize the attached order."))
-	assert.True(t, strings.Contains(textContent, `"orderId":"ABC-123"`))
-	assert.True(t, strings.Contains(textContent, `"total":4299`))
-	assert.True(t, hasDocument)
-	assert.Equal(t, docURI, "https://example.com/invoice.pdf")
-}
-
-// Messages that hold parts but carry no renderable content are rejected
-// with InvalidParams, not InternalError.
-func TestServerEmptyPartsRejectedAsInvalidParams(t *testing.T) {
-	model := &fakeLLM{generate: func(ctx context.Context, opts ...llm.Option) (*llm.Response, error) {
-		return textResponse("ok"), nil
-	}}
-	agent := buildAgent(t, model)
-	server, err := a2a.NewServer(a2a.ServerOptions{Agent: agent})
-	assert.NoError(t, err)
-	ts := httptest.NewServer(server.Handler())
-	defer ts.Close()
-
-	client, err := a2a.NewClient(a2a.ClientOptions{Endpoint: ts.URL + "/"})
-	assert.NoError(t, err)
-
-	_, err = client.SendMessage(context.Background(), &a2a.Message{
-		MessageID: "m1",
-		Role:      a2a.RoleUser,
-		// A data part with an empty map is effectively empty.
-		Parts: []a2a.Part{a2a.NewDataPart(nil)},
-	}, nil)
-	assert.Error(t, err)
-	rpcErr, ok := err.(*a2a.RPCError)
-	assert.True(t, ok)
-	assert.Equal(t, rpcErr.Code, a2a.ErrorCodeInvalidParams)
-}
-
-// ---------------------------------------------------------------------------
-// Server: resume does not duplicate the incoming user message in history
-// ---------------------------------------------------------------------------
-
-func TestServerResumeDoesNotDuplicateUserMessage(t *testing.T) {
-	var callNum atomic.Int32
-	model := &fakeLLM{generate: func(ctx context.Context, opts ...llm.Option) (*llm.Response, error) {
-		n := callNum.Add(1)
-		if n == 1 {
-			return toolCallResponse("ask", "call_1"), nil
-		}
-		return textResponse("all set"), nil
-	}}
-	agent := buildAgent(t, model, &suspendingTool{})
-	server, err := a2a.NewServer(a2a.ServerOptions{Agent: agent})
-	assert.NoError(t, err)
-	ts := httptest.NewServer(server.Handler())
-	defer ts.Close()
-
-	client, err := a2a.NewClient(a2a.ClientOptions{Endpoint: ts.URL + "/"})
-	assert.NoError(t, err)
-
-	task, err := client.SendMessage(context.Background(), &a2a.Message{
-		MessageID: "m1",
-		Role:      a2a.RoleUser,
-		Parts:     []a2a.Part{a2a.NewTextPart("kick off")},
-	}, nil)
-	assert.NoError(t, err)
-	assert.Equal(t, task.Status.State, a2a.TaskStateInputRequired)
-
-	resumed, err := client.SendMessage(context.Background(), &a2a.Message{
-		MessageID: "m2",
-		Role:      a2a.RoleUser,
-		Parts:     []a2a.Part{a2a.NewTextPart("approved")},
-		TaskID:    task.ID,
-		ContextID: task.ContextID,
-	}, nil)
-	assert.NoError(t, err)
-	assert.Equal(t, resumed.Status.State, a2a.TaskStateCompleted)
-
-	seen := 0
-	for _, h := range resumed.History {
-		if h.MessageID == "m2" {
-			seen++
-		}
-	}
-	assert.Equal(t, seen, 1)
-}
-
-// ---------------------------------------------------------------------------
-// Client: bare Message result on message/send is wrapped into a Task
-// ---------------------------------------------------------------------------
-
-func TestClientSendMessageHandlesBareMessageResult(t *testing.T) {
-	// Stand up a minimal hand-rolled JSON-RPC server that returns a
-	// Message result rather than a Task. This is the spec-allowed shape
-	// our own server does not emit, so we verify the client-side
-	// adapter explicitly.
-	mux := http.NewServeMux()
-	mux.HandleFunc("/", func(w http.ResponseWriter, r *http.Request) {
-		var req struct {
-			ID     json.RawMessage `json:"id"`
-			Method string          `json:"method"`
-		}
-		assert.NoError(t, json.NewDecoder(r.Body).Decode(&req))
-		assert.Equal(t, req.Method, a2a.MethodMessageSend)
-		result := a2a.Message{
-			MessageID: "srv-1",
-			Role:      a2a.RoleAgent,
-			ContextID: "ctx-xyz",
-			Parts:     []a2a.Part{a2a.NewTextPart("direct reply")},
-		}
-		w.Header().Set("Content-Type", "application/json")
-		_ = json.NewEncoder(w).Encode(map[string]any{
-			"jsonrpc": "2.0",
-			"id":      req.ID,
-			"result":  result,
-		})
-	})
-	ts := httptest.NewServer(mux)
-	defer ts.Close()
-
-	client, err := a2a.NewClient(a2a.ClientOptions{Endpoint: ts.URL + "/"})
-	assert.NoError(t, err)
-
-	task, err := client.SendMessage(context.Background(), &a2a.Message{
-		MessageID: "m1",
-		Role:      a2a.RoleUser,
-		Parts:     []a2a.Part{a2a.NewTextPart("hi")},
-	}, nil)
-	assert.NoError(t, err)
-	assert.Equal(t, task.Status.State, a2a.TaskStateCompleted)
-	assert.Equal(t, task.ContextID, "ctx-xyz")
-	assert.Equal(t, a2a.ResponseText(task), "direct reply")
-	if synthetic, _ := task.Metadata["a2a.syntheticFromMessage"].(bool); !synthetic {
-		t.Fatalf("expected a2a.syntheticFromMessage=true, got metadata=%v", task.Metadata)
-	}
-}
-
-// ---------------------------------------------------------------------------
-// Cancellation propagation: cancel interrupts an in-flight turn
-// ---------------------------------------------------------------------------
-
-func TestServerCancelInterruptsInflightTurn(t *testing.T) {
-	model := &blockingLLM{
-		started: make(chan struct{}),
-		release: make(chan struct{}),
-		result:  textResponse("should not arrive"),
-	}
-	agent := buildAgent(t, model)
-
-	// Use a spy store that notifies when a task is first stored so we
-	// can learn the server-assigned task ID while the turn is in flight.
-	spy := &spyTaskStore{inner: a2a.NewMemoryTaskStore()}
-	server, err := a2a.NewServer(a2a.ServerOptions{
-		Agent:     agent,
-		TaskStore: spy,
-	})
-	assert.NoError(t, err)
-	ts := httptest.NewServer(server.Handler())
-	defer ts.Close()
-
-	client, err := a2a.NewClient(a2a.ClientOptions{Endpoint: ts.URL + "/"})
-	assert.NoError(t, err)
-
-	type result struct {
-		task *a2a.Task
-		err  error
-	}
-	ch := make(chan result, 1)
-	go func() {
-		task, err := client.SendMessage(context.Background(), &a2a.Message{
-			MessageID: "m1",
-			Role:      a2a.RoleUser,
-			Parts:     []a2a.Part{a2a.NewTextPart("do something slow")},
-		}, nil)
-		ch <- result{task, err}
-	}()
-
-	// Wait for the LLM to start — the turn is now in flight.
-	<-model.started
-
-	// The in-flight cancel is keyed by the task ID that runTurn generated
-	// before calling CreateResponse. We can't observe that ID from outside
-	// without a hook, so we verify the simpler contract: cancelling the
-	// HTTP request context (via closing the test server) propagates.
-	// Instead, just release the LLM and verify normal completion; the
-	// real cancellation test is TestServerCancelBeforeResume below and
-	// the concurrency tests.
-	close(model.release)
-	r := <-ch
-	assert.NoError(t, r.err)
-	assert.Equal(t, r.task.Status.State, a2a.TaskStateCompleted)
-}
-
-// spyTaskStore wraps a TaskStore and exposes List for test assertions.
-type spyTaskStore struct {
-	inner a2a.TaskStore
-}
-
-func (s *spyTaskStore) Put(ctx context.Context, rec *a2a.TaskRecord) error {
-	return s.inner.Put(ctx, rec)
-}
-func (s *spyTaskStore) Get(ctx context.Context, id string) (*a2a.TaskRecord, bool, error) {
-	return s.inner.Get(ctx, id)
-}
-func (s *spyTaskStore) Delete(ctx context.Context, id string) error {
-	return s.inner.Delete(ctx, id)
-}
-func (s *spyTaskStore) List(ctx context.Context) ([]*a2a.TaskRecord, error) {
-	return s.inner.List(ctx)
-}
-
-// TestServerCancelBeforeResume verifies that cancelling a suspended task
-// prevents resume from succeeding.
-func TestServerCancelBeforeResume(t *testing.T) {
-	var callNum atomic.Int32
-	model := &fakeLLM{generate: func(ctx context.Context, opts ...llm.Option) (*llm.Response, error) {
-		n := callNum.Add(1)
-		if n == 1 {
-			return toolCallResponse("ask", "call_1"), nil
-		}
-		return textResponse("done"), nil
-	}}
-	agent := buildAgent(t, model, &suspendingTool{})
-	server, err := a2a.NewServer(a2a.ServerOptions{Agent: agent})
-	assert.NoError(t, err)
-	ts := httptest.NewServer(server.Handler())
-	defer ts.Close()
-
-	client, err := a2a.NewClient(a2a.ClientOptions{Endpoint: ts.URL + "/"})
-	assert.NoError(t, err)
-
-	task, err := client.SendMessage(context.Background(), &a2a.Message{
-		MessageID: "m1",
-		Role:      a2a.RoleUser,
-		Parts:     []a2a.Part{a2a.NewTextPart("start")},
-	}, nil)
-	assert.NoError(t, err)
-	assert.Equal(t, task.Status.State, a2a.TaskStateInputRequired)
-
-	_, err = client.CancelTask(context.Background(), task.ID)
-	assert.NoError(t, err)
-
-	// Attempting to resume a canceled task should fail.
-	_, err = client.SendMessage(context.Background(), &a2a.Message{
-		MessageID: "m2",
-		Role:      a2a.RoleUser,
-		Parts:     []a2a.Part{a2a.NewTextPart("too late")},
-		TaskID:    task.ID,
-	}, nil)
-	assert.Error(t, err)
-}
-
-// ---------------------------------------------------------------------------
-// TaskStore.List
-// ---------------------------------------------------------------------------
-
-func TestMemoryTaskStoreList(t *testing.T) {
-	store := a2a.NewMemoryTaskStore()
-	ctx := context.Background()
-
-	recs, err := store.List(ctx)
-	assert.NoError(t, err)
-	assert.Len(t, recs, 0)
-
-	_ = store.Put(ctx, &a2a.TaskRecord{Task: &a2a.Task{ID: "t1", Status: a2a.TaskStatus{State: a2a.TaskStateCompleted}}})
-	_ = store.Put(ctx, &a2a.TaskRecord{Task: &a2a.Task{ID: "t2", Status: a2a.TaskStatus{State: a2a.TaskStateWorking}}})
-	_ = store.Put(ctx, &a2a.TaskRecord{Task: &a2a.Task{ID: "t3", Status: a2a.TaskStatus{State: a2a.TaskStateCompleted}}})
-
-	recs, err = store.List(ctx)
-	assert.NoError(t, err)
-	assert.Len(t, recs, 3)
-
-	_ = store.Delete(ctx, "t2")
-	recs, err = store.List(ctx)
-	assert.NoError(t, err)
-	assert.Len(t, recs, 2)
-}
-
-// ---------------------------------------------------------------------------
-// Non-text content projection: image content → file part in artifacts
-// ---------------------------------------------------------------------------
-
-func TestServerNonTextContentInArtifacts(t *testing.T) {
-	model := &fakeLLM{generate: func(ctx context.Context, opts ...llm.Option) (*llm.Response, error) {
-		return imageResponse("Here is the chart"), nil
-	}}
-	agent := buildAgent(t, model)
-	server, err := a2a.NewServer(a2a.ServerOptions{Agent: agent})
-	assert.NoError(t, err)
-	ts := httptest.NewServer(server.Handler())
-	defer ts.Close()
-
-	client, err := a2a.NewClient(a2a.ClientOptions{Endpoint: ts.URL + "/"})
-	assert.NoError(t, err)
-
-	task, err := client.SendMessage(context.Background(), &a2a.Message{
-		MessageID: "m1",
-		Role:      a2a.RoleUser,
-		Parts:     []a2a.Part{a2a.NewTextPart("Make me a chart")},
-	}, nil)
-	assert.NoError(t, err)
-	assert.Equal(t, task.Status.State, a2a.TaskStateCompleted)
-	assert.True(t, len(task.Artifacts) > 0)
-
-	art := task.Artifacts[0]
-	assert.Len(t, art.Parts, 2)
-	assert.True(t, art.Parts[0].IsText())
-	assert.Equal(t, art.Parts[0].Text, "Here is the chart")
-	assert.True(t, art.Parts[1].IsURL())
-	assert.Equal(t, art.Parts[1].MediaType, "image/png")
-	assert.Equal(t, art.Parts[1].URL, "https://example.com/chart.png")
-
-	// History should also carry the full content.
-	var agentMsgs []*a2a.Message
-	for _, h := range task.History {
-		if h.Role == a2a.RoleAgent {
-			agentMsgs = append(agentMsgs, h)
-		}
-	}
-	assert.True(t, len(agentMsgs) > 0)
-	assert.Len(t, agentMsgs[0].Parts, 2)
-}
-
-// ---------------------------------------------------------------------------
-// Streaming emits all artifacts
-// ---------------------------------------------------------------------------
-
-func TestServerStreamEmitsAllArtifacts(t *testing.T) {
-	model := &fakeLLM{generate: func(ctx context.Context, opts ...llm.Option) (*llm.Response, error) {
-		return imageResponse("chart description"), nil
-	}}
-	agent := buildAgent(t, model)
-	server, err := a2a.NewServer(a2a.ServerOptions{Agent: agent})
-	assert.NoError(t, err)
-	ts := httptest.NewServer(server.Handler())
-	defer ts.Close()
-
-	client, err := a2a.NewClient(a2a.ClientOptions{Endpoint: ts.URL + "/"})
-	assert.NoError(t, err)
-
-	var artifactEvents []*a2a.TaskArtifactUpdateEvent
-	var statusEvents []*a2a.TaskStatusUpdateEvent
-	err = client.StreamMessage(context.Background(), &a2a.Message{
-		MessageID: "m1",
-		Role:      a2a.RoleUser,
-		Parts:     []a2a.Part{a2a.NewTextPart("stream chart")},
-	}, nil, func(ev *a2a.StreamEvent) error {
-		if ev.ArtifactUpdate != nil {
-			artifactEvents = append(artifactEvents, ev.ArtifactUpdate)
-		}
-		if ev.StatusUpdate != nil {
-			statusEvents = append(statusEvents, ev.StatusUpdate)
-		}
-		return nil
-	})
-	assert.NoError(t, err)
-
-	// Should have artifact event(s) with both text and image parts.
-	assert.True(t, len(artifactEvents) > 0)
-	art := artifactEvents[0]
-	assert.Len(t, art.Artifact.Parts, 2)
-	assert.True(t, art.Artifact.Parts[0].IsText())
-	assert.True(t, art.Artifact.Parts[1].IsURL())
-
-	// Final status should be completed.
-	final := statusEvents[len(statusEvents)-1]
-	assert.Equal(t, final.Status.State, a2a.TaskStateCompleted)
-}
-
-// ---------------------------------------------------------------------------
-// Multi-pending-tool-call resume with structured DataPart
-// ---------------------------------------------------------------------------
-
-func TestServerMultiPendingCallResumeWithDataPart(t *testing.T) {
-	var callNum atomic.Int32
-	model := &fakeLLM{generate: func(ctx context.Context, opts ...llm.Option) (*llm.Response, error) {
-		n := callNum.Add(1)
-		if n == 1 {
-			return multiToolCallResponse(
-				struct{ name, id string }{"ask", "call_1"},
-				struct{ name, id string }{"confirm", "call_2"},
-			), nil
-		}
-		return textResponse("Both answered, thanks."), nil
-	}}
-	agent := buildParallelAgent(t, model, &suspendingTool{}, &confirmTool{})
-	server, err := a2a.NewServer(a2a.ServerOptions{Agent: agent})
-	assert.NoError(t, err)
-	ts := httptest.NewServer(server.Handler())
-	defer ts.Close()
-
-	client, err := a2a.NewClient(a2a.ClientOptions{Endpoint: ts.URL + "/"})
-	assert.NoError(t, err)
-
-	task, err := client.SendMessage(context.Background(), &a2a.Message{
-		MessageID: "m1",
-		Role:      a2a.RoleUser,
-		Parts:     []a2a.Part{a2a.NewTextPart("parallel request")},
-	}, nil)
-	assert.NoError(t, err)
-	assert.Equal(t, task.Status.State, a2a.TaskStateInputRequired)
-
-	// Resume with structured toolResults DataPart.
-	resumed, err := client.SendMessage(context.Background(), &a2a.Message{
-		MessageID: "m2",
-		Role:      a2a.RoleUser,
-		Parts: []a2a.Part{
-			a2a.NewTextPart("see attached"),
-			a2a.NewDataPart(map[string]any{
-				"toolResults": map[string]any{
-					"call_1": "approved",
-					"call_2": "confirmed",
-				},
-			}),
-		},
-		TaskID:    task.ID,
-		ContextID: task.ContextID,
-	}, nil)
-	assert.NoError(t, err)
-	assert.Equal(t, resumed.Status.State, a2a.TaskStateCompleted)
-	assert.Equal(t, a2a.ResponseText(resumed), "Both answered, thanks.")
-}
-
-// TestServerMultiPendingCallResumeWithTextBroadcast verifies that a plain
-// text resume broadcasts to all pending calls when no DataPart is present.
-func TestServerMultiPendingCallResumeWithTextBroadcast(t *testing.T) {
-	var callNum atomic.Int32
-	model := &fakeLLM{generate: func(ctx context.Context, opts ...llm.Option) (*llm.Response, error) {
-		n := callNum.Add(1)
-		if n == 1 {
-			return multiToolCallResponse(
-				struct{ name, id string }{"ask", "call_1"},
-				struct{ name, id string }{"confirm", "call_2"},
-			), nil
-		}
-		return textResponse("Got it."), nil
-	}}
-	agent := buildParallelAgent(t, model, &suspendingTool{}, &confirmTool{})
-	server, err := a2a.NewServer(a2a.ServerOptions{Agent: agent})
-	assert.NoError(t, err)
-	ts := httptest.NewServer(server.Handler())
-	defer ts.Close()
-
-	client, err := a2a.NewClient(a2a.ClientOptions{Endpoint: ts.URL + "/"})
-	assert.NoError(t, err)
-
-	task, err := client.SendMessage(context.Background(), &a2a.Message{
-		MessageID: "m1",
-		Role:      a2a.RoleUser,
-		Parts:     []a2a.Part{a2a.NewTextPart("go")},
-	}, nil)
-	assert.NoError(t, err)
-	assert.Equal(t, task.Status.State, a2a.TaskStateInputRequired)
-
-	// Resume with plain text — should broadcast to all pending calls.
-	resumed, err := client.SendMessage(context.Background(), &a2a.Message{
-		MessageID: "m2",
-		Role:      a2a.RoleUser,
-		Parts:     []a2a.Part{a2a.NewTextPart("yes to all")},
-		TaskID:    task.ID,
-		ContextID: task.ContextID,
-	}, nil)
-	assert.NoError(t, err)
-	assert.Equal(t, resumed.Status.State, a2a.TaskStateCompleted)
-}
-
-// ---------------------------------------------------------------------------
-// Concurrency: parallel requests on the same server
-// ---------------------------------------------------------------------------
-
-func TestServerConcurrentMessageSend(t *testing.T) {
-	var counter atomic.Int32
-	model := &fakeLLM{generate: func(ctx context.Context, opts ...llm.Option) (*llm.Response, error) {
-		n := counter.Add(1)
-		return textResponse(fmt.Sprintf("response %d", n)), nil
-	}}
-	agent := buildAgent(t, model)
-	server, err := a2a.NewServer(a2a.ServerOptions{Agent: agent})
-	assert.NoError(t, err)
-	ts := httptest.NewServer(server.Handler())
-	defer ts.Close()
-
-	client, err := a2a.NewClient(a2a.ClientOptions{Endpoint: ts.URL + "/"})
-	assert.NoError(t, err)
-
-	const N = 20
-	var wg sync.WaitGroup
-	errs := make([]error, N)
-	tasks := make([]*a2a.Task, N)
-
-	for i := 0; i < N; i++ {
-		wg.Add(1)
-		go func(idx int) {
-			defer wg.Done()
-			task, err := client.SendMessage(context.Background(), &a2a.Message{
-				MessageID: fmt.Sprintf("m%d", idx),
-				Role:      a2a.RoleUser,
-				Parts:     []a2a.Part{a2a.NewTextPart(fmt.Sprintf("request %d", idx))},
-			}, nil)
-			tasks[idx] = task
-			errs[idx] = err
-		}(i)
-	}
-	wg.Wait()
-
-	for i := 0; i < N; i++ {
-		assert.NoError(t, errs[i])
-		assert.Equal(t, tasks[i].Status.State, a2a.TaskStateCompleted)
-		assert.NotEqual(t, tasks[i].ID, "")
-	}
-
-	// All task IDs should be unique.
-	ids := make(map[string]bool)
-	for _, task := range tasks {
-		assert.False(t, ids[task.ID])
-		ids[task.ID] = true
-	}
-}
-
-func TestServerConcurrentSendAndGet(t *testing.T) {
-	var callNum atomic.Int32
-	model := &fakeLLM{generate: func(ctx context.Context, opts ...llm.Option) (*llm.Response, error) {
-		n := callNum.Add(1)
-		if n == 1 {
-			return toolCallResponse("ask", "call_1"), nil
-		}
-		return textResponse("done"), nil
-	}}
-	agent := buildAgent(t, model, &suspendingTool{})
-	server, err := a2a.NewServer(a2a.ServerOptions{Agent: agent})
-	assert.NoError(t, err)
-	ts := httptest.NewServer(server.Handler())
-	defer ts.Close()
-
-	client, err := a2a.NewClient(a2a.ClientOptions{Endpoint: ts.URL + "/"})
-	assert.NoError(t, err)
-
-	// Create a suspended task.
-	task, err := client.SendMessage(context.Background(), &a2a.Message{
-		MessageID: "m1",
-		Role:      a2a.RoleUser,
-		Parts:     []a2a.Part{a2a.NewTextPart("start")},
-	}, nil)
-	assert.NoError(t, err)
-
-	// Hammer tasks/get concurrently while the task exists.
-	const N = 10
-	var wg sync.WaitGroup
-	for i := 0; i < N; i++ {
-		wg.Add(1)
-		go func() {
-			defer wg.Done()
-			got, err := client.GetTask(context.Background(), task.ID)
-			assert.NoError(t, err)
-			assert.Equal(t, got.ID, task.ID)
-		}()
-	}
-	wg.Wait()
-}
-
-// ---------------------------------------------------------------------------
-// SuspendReason: auth_required maps to A2A auth-required state
-// ---------------------------------------------------------------------------
-
-type authSuspendTool struct{}
-
-func (t *authSuspendTool) Name() string                        { return "auth_gate" }
-func (t *authSuspendTool) Description() string                 { return "Require auth" }
-func (t *authSuspendTool) Schema() *dive.Schema                { return nil }
-func (t *authSuspendTool) Annotations() *dive.ToolAnnotations  { return &dive.ToolAnnotations{Title: "AuthGate"} }
-func (t *authSuspendTool) Call(ctx context.Context, input any) (*dive.ToolResult, error) {
-	return dive.NewSuspendResultWithReason("Sign in to continue",
-		dive.SuspendReasonAuth, map[string]any{"auth_url": "https://example.com/oauth"}), nil
-}
-
-func TestServerAuthRequiredSuspendMapsToAuthState(t *testing.T) {
-	var callNum atomic.Int32
-	model := &fakeLLM{generate: func(ctx context.Context, opts ...llm.Option) (*llm.Response, error) {
-		n := callNum.Add(1)
-		if n == 1 {
-			return toolCallResponse("auth_gate", "call_auth"), nil
-		}
-		return textResponse("authenticated"), nil
-	}}
-	agent := buildAgent(t, model, &authSuspendTool{})
-	server, err := a2a.NewServer(a2a.ServerOptions{Agent: agent})
-	assert.NoError(t, err)
-	ts := httptest.NewServer(server.Handler())
-	defer ts.Close()
-
-	client, err := a2a.NewClient(a2a.ClientOptions{Endpoint: ts.URL + "/"})
-	assert.NoError(t, err)
-
-	task, err := client.SendMessage(context.Background(), &a2a.Message{
-		MessageID: "m1",
-		Role:      a2a.RoleUser,
-		Parts:     []a2a.Part{a2a.NewTextPart("access the resource")},
-	}, nil)
-	assert.NoError(t, err)
-	assert.Equal(t, task.Status.State, a2a.TaskStateAuthRequired)
-	assert.True(t, strings.Contains(task.Status.Message.TextContent(), "Sign in"))
-
-	// Resume with auth result.
-	resumed, err := client.SendMessage(context.Background(), &a2a.Message{
-		MessageID: "m2",
-		Role:      a2a.RoleUser,
-		TaskID:    task.ID,
-		Parts:     []a2a.Part{a2a.NewTextPart("token_abc")},
-	}, nil)
-	assert.NoError(t, err)
-	assert.Equal(t, resumed.Status.State, a2a.TaskStateCompleted)
-}
-
-// ---------------------------------------------------------------------------
-// Image file parts project to ImageContent, not DocumentContent
-// ---------------------------------------------------------------------------
-
-func TestServerImagePartProjectsToImageContent(t *testing.T) {
-	var hasImage bool
-	var imageURI string
-	model := &fakeLLM{generate: func(ctx context.Context, opts ...llm.Option) (*llm.Response, error) {
-		cfg := &llm.Config{}
-		cfg.Apply(opts...)
-		for _, msg := range cfg.Messages {
-			if msg.Role != llm.User {
-				continue
-			}
-			for _, c := range msg.Content {
-				if img, ok := c.(*llm.ImageContent); ok {
-					hasImage = true
-					if img.Source != nil {
-						imageURI = img.Source.URL
-					}
-				}
-			}
-		}
-		return textResponse("nice picture"), nil
-	}}
-	agent := buildAgent(t, model)
-	server, err := a2a.NewServer(a2a.ServerOptions{Agent: agent})
-	assert.NoError(t, err)
-	ts := httptest.NewServer(server.Handler())
-	defer ts.Close()
-
-	client, err := a2a.NewClient(a2a.ClientOptions{Endpoint: ts.URL + "/"})
-	assert.NoError(t, err)
-
-	task, err := client.SendMessage(context.Background(), &a2a.Message{
-		MessageID: "m1",
-		Role:      a2a.RoleUser,
-		Parts: []a2a.Part{
-			a2a.NewTextPart("What's in this image?"),
-			{URL: "https://example.com/photo.png", MediaType: "image/png", Filename: "photo.png"},
-		},
-	}, nil)
-	assert.NoError(t, err)
-	assert.Equal(t, task.Status.State, a2a.TaskStateCompleted)
-	assert.True(t, hasImage)
-	assert.Equal(t, imageURI, "https://example.com/photo.png")
-}
