+++ go.mod
@@ -3,6 +3,7 @@ module github.com/deepnoodle-ai/dive
 go 1.25.0
 
 require (
+	github.com/a2aproject/a2a-go/v2 v2.2.0
 	github.com/deepnoodle-ai/wonton v0.0.29
 	github.com/gobwas/glob v0.2.3
 	github.com/google/uuid v1.6.0
@@ -14,7 +15,9 @@ require (
 	github.com/google/go-cmp v0.7.0 // indirect
 	github.com/kr/pretty v0.3.1 // indirect
 	github.com/rogpeppe/go-internal v1.14.1 // indirect
+	golang.org/x/mod v0.33.0 // indirect
 	golang.org/x/net v0.51.0 // indirect
+	golang.org/x/sync v0.19.0 // indirect
 	golang.org/x/sys v0.41.0 // indirect
 	gopkg.in/check.v1 v1.0.0-20201130134442-10cb98267c6c // indirect
 )
